IOUtils Changes
- added new NormalizeFolder method to format the path parameter, so that it is formed correctly.
If the path is a local root path (eg. d:) then it must end with \\.
If the path is not a root path, or is a Unc path, then any ending \\ must be removed.

This fix was due to an issue with the previous code, as it kept appending a \ to the end when the path was a Unc path.